feat: move context managers to top level class
